# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "HealthPass Data Submission API",
    "description": "Swagger definition for HealthPass data submission API",
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "/api/v1/datasubmission",
      "description": "CDT"
    }
  ],
  "paths": {
...
    "/register-code/{regEntity}/upload": {
      "post": {
        "tags": [
          "register-code"
        ],
        "summary": "Upload registration codes",
        "description": "Upload registration codes for specified entity",
        "operationId": "uploadRegistrationCodes",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "regEntity",
            "in": "path",
            "description": "Registration entity",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "expiresAt",
            "in": "query",
            "description": "Expiration date (seconds since Unix epoch)",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "expiresIn",
            "in": "query",
            "description": "Registration code lifespan (seconds)",
            "required": false,
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "description": "User registration codes",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationCodesRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully uploaded registration codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationCodesResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/AllUploadedRegistrationCodesError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "409": {
            "$ref": "#/components/responses/SomeUploadedRegistrationCodesError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/register-code/{regEntity}/generate/{numCodes}": {
      "post": {
        "tags": [
          "register-code"
        ],
        "summary": "Generate registration codes",
        "description": "Generate registration codes for specified entity",
        "operationId": "generateRegistrationCodes",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "regEntity",
            "in": "path",
            "description": "Registration entity",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "numCodes",
            "in": "path",
            "description": "Number of registration codes",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "expiresAt",
            "in": "query",
            "description": "Expiration date (seconds since Unix epoch)",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "expiresIn",
            "in": "query",
            "description": "Registration code lifespan (seconds)",
            "required": false,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successfully generated registration codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationCodesResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/AllGeneratedRegistrationCodesError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "409": {
            "$ref": "#/components/responses/SomeGeneratedRegistrationCodesError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "RegistrationCodesRequest": {
        "type": "object",
        "required": [
          "registrationCodes"
        ],
        "properties": {
          "registrationCodes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "RegistrationCodesResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "docs": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "registerCode": {
                  "type": "string"
                },
                "expirationTimestamp": {
                  "type": "string"
                },
                "createdTimestamp": {
                  "type": "string"
                },
                "updatedTimestamp": {
                  "type": "string"
                },
                "status": {
                  "type": "string"
                }
              }
            }
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "reason": {
                  "type": "string"
                },
                "registerCode": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
      "BasicErrorMessage": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
...
    },
    "responses": {
...
      "UnauthorizedError": {
        "description": "Logged in user is unauthorized to make this request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BasicErrorMessage"
            }
          }
        }
      },
...
      "ServerError": {
        "description": "Unexpected server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BasicErrorMessage"
            }
          }
        }
      },
...
      "AllUploadedRegistrationCodesError": {
        "description": "Unable to upload any registration codes",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/RegistrationCodesResponse"
            }
          }
        }
      },
      "SomeUploadedRegistrationCodesError": {
        "description": "Unable to upload some registration codes",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/RegistrationCodesResponse"
            }
          }
        }
      },
      "AllGeneratedRegistrationCodesError": {
        "description": "Unable to generate any registration codes",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/RegistrationCodesResponse"
            }
          }
        }
      },
      "SomeGeneratedRegistrationCodesError": {
        "description": "Unable to generate some registration codes",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/RegistrationCodesResponse"
            }
          }
        }
      }
    }
  }
}